tests and more stuff

+++ db.rb
@@ -23,11 +23,12 @@ module KurzyDB
             String  :url, :unique => false, :empty => false
             Integer :counter, default: 0
             DateTime	:timestamp, default: Sequel::CURRENT_TIMESTAMP
+            Boolean :private, default: 1
         end
         create_table unless table_exists?
     end
 
-    def KurzyDB.add(url:, short:nil)
+    def KurzyDB.add(url:, short:nil, priv:true)
         return unless url
 
         s = short
@@ -35,22 +36,26 @@ module KurzyDB
             s = KurzyDB.gen_hash()
         end
         begin
-            KURL.insert(url: url, short: s)
+            KURL.insert(url: url, short: s, private: priv)
         rescue Sequel::UniqueConstraintViolation
             if s == short
                 raise KurzyDB::Error.new("The short url #{short} already exists")
             else
                 s = KurzyDB.gen_hash()
-                KurzyDB.insert(url: url, short: s)
+                KurzyDB.insert(url: url, short: s, private: priv)
             end
         end
         return s
     end
 
     def KurzyDB.delete(short:)
-        row = KurzyDB.where(short: short)
-        row.delete()
-        return row.to_hash
+        row = KURL.where(short: short).first
+        if row
+            row.delete()
+            return row.to_hash
+        else
+            raise KurzyDB::Error.new("The short url #{short} doesn't exist")
+        end
     end
 
     def KurzyDB.gen_hash(length=6)
@@ -58,13 +63,22 @@ module KurzyDB
     end
 
     def KurzyDB.get_url(short:)
-        $stderr.puts KURL.where(short: short).sql
         res =  KURL.where(short: short).first()
         if res
+            KURL.where(id: res[:id]).update(counter: res[:counter]+1)
             return res[:url]
         else
             return nil
         end
+    end
+
+    def KurzyDB.list(max:nil, priv: false)
+        rows = priv ? KURL.all :  KURL.where(private: false)
+        return rows.map{|row| row.to_hash}
+    end
+
+    def KurzyDB.truncate()
+        KURL.truncate()
     end
 end
